requirements.txt
streamlit
pandas
python-barcode
pillow
pyzbar
numpy
qrcode
uuid
openpyxl # Untuk ekspor ke Excel
plotly # Untuk Grafik Interaktif
shutil # Untuk backup
zipfile # Untuk kompresi